댓글 등록, 삭제 구현

+++ BACKEND/RECIPICK-PROJECT/src/main/java/SITE/RECIPICK/RECIPICK_PROJECT/dto/SearchPostDto.java
@@ -1,7 +1,6 @@
 package SITE.RECIPICK.RECIPICK_PROJECT.dto;
 
 import java.time.LocalDateTime;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -9,17 +8,18 @@ import lombok.Data;
 @Data
 public class SearchPostDto {
 
-    private Long postId;
-    private Integer userId;
-    private String title;
-    private String foodName;
-    private String rcpImgUrl;
-    private String rcpStepsImg;
-    private Integer viewCount;
-    private Integer likeCount;
-    private LocalDateTime createdAt;
-    private Integer subScore; // 선택적 계산용 필드
+  private Integer postId;
+  private Integer userId;
+  private String title;
+  private String foodName;
+  private String rcpImgUrl;
+  private String rcpStepsImg;
+  private Integer viewCount;
+  private Integer likeCount;
+  private LocalDateTime createdAt;
+  private Integer subScore; // 선택적 계산용 필드
 
-    // 기본 생성자
-    public SearchPostDto() {}
+  // 기본 생성자
+  public SearchPostDto() {
+  }
 }
